requirements.txt
pandas~=1.3.3
numpy~=1.19.5
sklearn~=0.0
pyquery~=1.4.3
swifter~=1.0.9
python-whois
waybackpy~=2.4.4
shodan~=1.25.0
matplotlib~=3.4.3
seaborn~=0.11.2
xgboost~=1.5.0
scikit-learn~=1.0
requests~=2.26.0